+++ RELEASE-NOTES.md
@@ -39,6 +39,8 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Inlet: Added support for deeply nested containers of structs
 - Inlet: Added support for `void` and strings in Lua-defined functions
 - Inlet: Added `get<std::vector<T>>` for retrieving arrays without index information
+- Inlet: Added a new `Writer` for generating JSON schemas which can be used by text editors
+  for autocompletion
 - Inlet: SphinxWriter will now document the signature of function callbacks added to a schema
 
 ### Changed